+++ src/Examples/BuildAndParseExample.cs
@@ -13,8 +13,8 @@ public static class BuildAndParseExample
 
         MimeTypeInfo mimeType1 = 
             MimeTypeBuilder.Create("application", "x-stuff")
-                           .AppendParameter("first-parameter", longParameterValue, "en")
-                           .AppendParameter("second-parameter", "Parameter with  \\, = and \".")
+                           .AddParameter("first-parameter", longParameterValue, "en")
+                           .AddParameter("second-parameter", "Parameter with  \\, = and \".")
                            .Build();
 
         string s = mimeType1.ToString(MimeFormats.LineWrapping);